Add `IRInspector` protocol to allow easy inspecting of IR
Both `Module` and `IRWriter` needed functionality to inspect a function.
We don't want to duplicate functionality.

Changes:
- add `IRInspector` protocol that needs to obtain a `Function` from its ID
- move the functionality to inspect the IR to `IRInspector`
- make `Module` extend `IRInspector` and remove inspection functions from it
- make `IRWriter` extend `IRInspector` and remove inspection functions from it
- changed `DominatorTree` to take `Function` instead of `Module`
