+++ Sources/IR/Analysis/DominatorTree.swift
@@ -22,12 +22,13 @@ struct DominatorTree {
   /// The immediate dominators of each basic block.
   private var immediateDominators: [Node: Node?]
 
-  /// Creates the dominator tree of `f`, which is in `m`, using the given `cfg`.
-  init(function f: Function.ID, cfg: ControlFlowGraph, in m: Module) {
+  /// Creates the dominator tree of `ir`using the given `cfg`.
+  init(function ir: Function, cfg: ControlFlowGraph? = nil) {
     // The following is an implementation of Cooper et al.'s fast dominance iterative algorithm
     // (see "A Simple, Fast Dominance Algorithm", 2001). First, build any spanning tree rooted at
     // the function's entry.
-    var t = SpanningTree(of: cfg, rootedAt: m[f].entry!)
+    let cfg2 = cfg ?? ir.cfg()
+    var t = SpanningTree(of: cfg2, rootedAt: ir.entry!)
 
     // Then, until a fixed point is reached, for each block `v` that has a predecessor `u` that
     // isn't `v`'s parent in the tree, assign `v`'s parent to the least common ancestor of `u` and
@@ -35,8 +36,8 @@ struct DominatorTree {
     var changed = true
     while changed {
       changed = false
-      for v in m[f].blocks.addresses {
-        for u in cfg.predecessors(of: v) where t.parent(v) != u {
+      for v in ir.blocks.addresses {
+        for u in cfg2.predecessors(of: v) where t.parent(v) != u {
           let lca = t.lowestCommonAncestor(u, t.parent(v)!)
           if lca != t.parent(v) {
             t.setParent(lca, forChild: v)
@@ -47,7 +48,7 @@ struct DominatorTree {
     }
 
     // The resulting tree encodes the immediate dominators.
-    root = m[f].entry!
+    root = ir.entry!
     immediateDominators = t.parents
   }
